lms runtime remove
