+++ src/subcommands/runtime/main.ts
@@ -1,5 +1,6 @@
 import { Command } from "@commander-js/extra-typings";
 import { ls } from "./list.js";
+import { remove } from "./remove.js";
 import { select } from "./select.js";
 
 // Create the runtime command
@@ -8,5 +9,6 @@ const runtimeCommand = new Command().name("runtime").description("Manage runtime
 // Add subcommands
 runtimeCommand.addCommand(ls);
 runtimeCommand.addCommand(select);
+runtimeCommand.addCommand(remove);
 
 export const runtime = runtimeCommand;
